cloud_functions/strava_webhook_handler/requirements.txt
-i https://pypi.org/simple
blinker==1.9.0; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
cloudevents==1.12.0
deprecation==2.1.0
flask==3.1.1; python_version >= '3.9'
functions-framework==3.8.3; python_version >= '3.5' and python_version < '4'
gunicorn==23.0.0; platform_system != 'Windows'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
python-decouple==3.8
watchdog==6.0.0; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
